chore: move the pipeline code to separate crate
This is needed to avoid circular dependencies.
