+++ dataplane/Cargo.toml
@@ -14,6 +14,7 @@ iptrie = { workspace = true }
 dyn-iter = { workspace = true }
 net = { workspace = true, features = ["serde"] }
 ordermap = { workspace = true, features = ["std"] }
+pipeline = { workspace = true }
 routing = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_yml = { workspace = true }
