+++ pipeline/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "pipeline"
+version = "0.1.0"
+edition = "2024"
+publish = false
+license = "Apache-2.0"
+
+[dependencies]
+dyn-iter = { workspace = true }
+id = { workspace = true }
+net = { workspace = true }
+ordermap = { workspace = true, features = ["std"] }
+thiserror = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+
+[dev-dependencies]
+net = { workspace = true, features = ["test_buffer"] }
